# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "chat_api",
      "contact": "garciavillegasjonathan@gmail.com",
      "description": "...",
      "version": "1.0"
    },
    "paths": {
      "/api/v1/users": {
        "post": {
          "tags": [
            "users"
          ],
          "summary": "register",
          "operationId": "register",
          "parameters": [],
          "requestBody": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/registerrequest"
                    },
                    {
                      "example": {
                        "firstname": "Shin",
                        "lastname": "Nouzen",
                        "email": "shin@gmail.com",
                        "password": "Greco1234*"
                      }
                    }
                  ]
                },
                "example": {
                  "firstname": "Shin",
                  "lastname": "Nouzen",
                  "email": "shin@gmail.com",
                  "password": "Greco1234*"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "deprecated": false
        },
        "get": {
            "security": [
                {
                  "bearer": []
                }
              ],
          "tags": [
            "users"
          ],
          "summary": "get all users",
          "operationId": "getallusers",
          "parameters": [],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "deprecated": false
        }
      },
...
      "/api/v1/users/{id}": {
        "put": {
          "tags": [
            "Misc"
          ],
          "summary": "upload avatar",
          "operationId": "uploadavatar",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32",
                "example": 2
              }
            }
          ],
          "requestBody": {
            "content": {
              "multipart/form-data": {
                "encoding": {},
                "schema": {
                  "required": [
                    "avatar"
                  ],
                  "type": "object",
                  "properties": {
                    "avatar": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            },
            "required": false
          },
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "deprecated": false,
          "security": [
            {
              "bearer": []
            }
          ]
        }
      }
    },
    "components": {
      "schemas": {
        "registerrequest": {
          "title": "registerrequest",
          "required": [
            "firstname",
            "lastname",
            "email",
            "password"
          ],
          "type": "object",
          "properties": {
            "firstname": {
              "type": "string"
            },
            "lastname": {
              "type": "string"
            },
            "email": {
              "type": "string"
            },
            "password": {
              "type": "string"
            }
          },
          "example": {
            "firstname": "Shin",
            "lastname": "Nouzen",
            "email": "shin@gmail.com",
            "password": "Greco1234*"
          }
        },
...
      },
      "securitySchemes": {
        "bearer": {
          "type": "http",
          "scheme": "bearer"
        }
      }
    },
    "security": [],
...
  }